+++ src/main/java/com/icebreaker/be/application/room/event/RoomStageEventListener.java
@@ -2,7 +2,7 @@ package com.icebreaker.be.application.room.event;
 
 
 import com.icebreaker.be.application.room.RoomStageService;
-import com.icebreaker.be.domain.EventPublisher;
+import com.icebreaker.be.domain.publisher.EventPublisher;
 import com.icebreaker.be.global.annotation.AsyncTransactionalEventListener;
 import com.icebreaker.be.infra.persistence.redis.message.PubSubMessage;
 import com.icebreaker.be.infra.persistence.redis.message.PubSubMessageType;
